fix inode recycling issue
Since commit 3b8fa1b420eb741de2ab99eeccb164461450ffdc each instance of
fakeroot handles correctly the filesystem inode recycling. However,
because each package is installed in its own instance of fakeroot, when
the fakeroot "databases" are concatenated, there may be duplicate due
to one package recycling the inode of a file from another package. So,
commit 3233e628107ccc5e59f01bb4845d21b85802792e adds a filter tool to
sip the fakeroot "databases" through the list of files we know we have
installed. This commit further replaces the global stripping phase
with a per-package stripping phase, allowing the per-package fakeroot
"databases" to reference the stripped files, instead of files prior to
stripping. The result is slower, but correct. For speed, stripping can
be disabled anyway.