toolchain: remove the lib32 and lib64 symlinks
They were only necessary for the loader to be found in the right place,
but this is no longer needed since it is now installed where the kernel
will look for it.

We have to keep the arm-linux-gnueabihf symlinks though, they seem to
be required for systems compiled with such toolchain to boot.