arm: rework VFP options
having to select soft or hard fp ABI in the configuration is error-prone,
as a toolchain only supports one or the other. The problem, however is that
some toolchain (CodeSourery in particular) will not generate hard floating
point code with only -mfpu=vfp, but require to pass -mfloat-abi=softfp as well.
To handle this case, get the toolchain package to generate a file with the
correct -mfloat-abi option for the current toolchain, and include it in
Makefile.pkgbuild to generate correct options when VFP is enabled.

At this change, rework the VFP options menu to allow selecting all the possible
values of the -mfpu option.