toolchain: a better fix for weird loaders issues
When ld.so.cache is missing, the glibc linux loader uses hardcoded
libraries search paths. For some toolchains, these hardcoded search paths
do not contain /lib /usr/lib. Known to have this issue:
- the Linaro armhf toolchain
- the Fedora x86_64 toolchain
- maybe the crosstool-ng generated ppc64 toolchain

Since we put the glibc libraries in /lib, this is a problem, so, to avoid
this situation generate /etc/ld.so.cache on the host.

Unfortunately, this is not enough, as the loader seems to look for libc.so.6
in the hardcoded path anyway, so workaround this by putting libc.so.6 in
/etc/ld.so.preload.