toolchain: fix weird loaders issue again
The previous fix for this issue:
e4919e1b23c293d79120346a84b5627a461a6fb8

relied on generating ld.so.cache on the build machine. Unfortunately
this only works if the build machine ldconfig is able to parse the
target machine binary format, on machines where it is not true,
ldconfig generates a valid, empty, ld.so.cache.

So, as a different workaround, extract the loader hard coded search
directory using the strings command, and create a symbolic link from
that directory to /lib. Also remove the generated ld.so.cache, as it
does more harm than good when the build machine generates an empty one.