introduce target package/logview
Allow opening logs in an editor. The editor is passed in the LOGVIEW
environment variable.

For instance, LOGVIEW could be $HOME/bin/logview.sh with
logview.sh the following script:

FILE=`pwd`/$1
if [ -n "$REMOTE" ]; then
    FILE="/$REMOTE:$FILE"
    EMACS="emacsclient -n"
else
    EMACS="emacs -nw"
fi

$EMACS --eval '(view-file "'$FILE'")' --eval '(compilation-mode)'

Allowing to open the logs in emacs and use emacs "compilation mode"
navigating from error to error with TAB and opening the corresponding
file with ENTER.