tftp: implement poka-yoke
Avoid making the copy of kernel, initramfs, and dtb to tftp separately
optional and with a configurable name. It is error-prone.

The copy to tftp directory is enabled or disabled globally, if it is
enabled and a file is generated, the file is copied to the tftp
directory, and the three file names are generated from a unique
directory name and base name as:
dirname/kernel-basename
dirname/dtb-basename
dirname/initramfs-basename