board/configuration handling reorganisation
This commit addresses the following issues:
1- the fact that we ended up setting "generic" as board name for all
boards configuration, and store the configuration files in a directory
under the boards/generic directory, simply because we wanted to share
the base files (those in boards/generic/skel);
2- the fact that the precious configuration files were in the build tree,
the one which should not contain anything precious, and were tedious to
save in the git repository.

In order to address the first issue, the following changes were made:
- the variable MKR_BOARD is renamed MKR_BOARD_NAME, the directory
boards/$(MKR_BOARD_NAME) contain some board specific files, namely:
- the mkrootfs configuration file, renamed from
boards/board_name_defconfig to boards/board_name/mkrootfs-defconfig;
- if MKR_BASE_SRC_BOARD is selected, the base files (in the skel
sub-directory), this option is off by default, which means that the base
files are taken in the boards/generic/skel directory;
- if MKR_LINUX_BOARD_CONFIG is selected, the Linux kernel configuration,
named linux-defconfig-$(KERNELVERSION), it is on by default, if it is off,
the name of a Linux defconfig target must be supplied;
- if MKR_BUSYBOX_BOARD_CONFIG is selected, the busybox configuration,
named busybox-defconfig, it is off by default, in which case the busybox
defconfig is used.

In order to address the second issue, a "saveconfig" target is added
which saves the various configuration files in the mkrootfs tree. In
order to "fork" a configuration:
- create a new build tree and call:
  make -C /path/to/mkrootfs O=/path/to/build-tree board_name_defconfig
- then run make menuconfig to change the MKR_BOARD_NAME variable
- then run make saveconfig to save the newly created board
configuration.

In order to recycle an old configuration, run "make oldconfig", be
careful when you answer the first question (it is the board name, it
should be a new directory name under mkrootfs boards directory), and you
can probably accept the other question default answers.